requirements.txt
Django>=5.0
psycopg2-binary>=2.9
djangorestframework
djangorestframework-simplejwt
python-dotenv
celery
redis
django-cors-headers
django-encrypted-model-fields